+++ Liersch.Json/Json/SLJsonNode.cs
@@ -20,6 +20,9 @@ namespace Liersch.Json
 {
   public sealed partial class SLJsonNode
   {
+    public static SLJsonNode Parse(string jsonExpression) { return SLJsonParser.Parse(jsonExpression); }
+
+
     public int Count
     {
       get